First Wave of Refactors
To compile on Linux, I need to refactor the codebase a bit
to allow Platform Windows to not be included on compilation.

Because the SubsystemFactory was doing it, I don't think
it would be possible without making these changes.

Also renamed many files, moved things around, and converted
space sizes to 2.